chore: Removed vite for now, only for building
