+++ packages/angular/src/builders/remote/schema.d.ts
@@ -1,23 +1,17 @@
 import type { JsonObject } from '@angular-devkit/core';
-import type { BuildNotificationOptions } from '@softarc/native-federation-runtime';
 import type { Plugin } from 'esbuild';
 import type { ApplicationBuilderOptions } from '@angular/build';
 
-export type NfSlimIndexOption = string | false | { input: string; output?: string };
-
-export interface NfSlimBuilderSchema extends JsonObject {
+export interface NfRemoteBuilderSchema extends JsonObject {
   tsConfig: string;
   dev: boolean;
-  port: number;
   rebuildDelay: number;
-  buildNotifications?: BuildNotificationOptions;
   watch: boolean;
   outputPath?: string;
   projectName?: string;
   verbose?: boolean;
   entryPoints?: string[];
   cacheExternalArtifacts?: boolean;
-  index?: NfSlimIndexOption;
 
   // Passthroughs to the Angular esbuild pipeline / asset copier.
   assets?: ApplicationBuilderOptions['assets'];
@@ -29,4 +23,4 @@ export interface NfSlimBuilderSchema extends JsonObject {
   preserveSymlinks?: ApplicationBuilderOptions['preserveSymlinks'];
 }
 
-export type NfSlimInternalOptions = { plugins: Plugin[] };
+export type NfRemoteInternalOptions = { plugins: Plugin[] };
