+++ packages/angular/src/builders/remote/schema.json
@@ -2,8 +2,8 @@
   "version": 2,
   "outputCapture": "direct-nodejs",
   "$schema": "http://json-schema.org/draft-07/schema",
-  "title": "native federation slim builder",
-  "description": "builder for native federation apps that only produces federation artifacts and serves them over a minimal Vite static server (no Angular application build)",
+  "title": "native federation remote builder",
+  "description": "builder for native federation remotes that only produces the federation artifacts (remoteEntry.json + referenced files) plus copied assets. No Angular application build and no dev server; serving is handled externally.",
   "type": "object",
   "properties": {
     "tsConfig": {
@@ -19,46 +19,13 @@
       "type": "boolean",
       "default": false
     },
-    "port": {
-      "type": "number",
-      "default": 0
-    },
     "entryPoints": {
       "type": "array"
-    },
-    "index": {
-      "description": "Configures the HTML file copied into the output directory. Accepts a string path to an HTML file, an object with explicit input/output paths, or `false` to skip emitting any index. When omitted, a minimal placeholder index.html is written.",
-      "oneOf": [
-        {
-          "type": "string"
-        },
-        {
-          "type": "object",
-          "additionalProperties": false,
-          "properties": {
-            "input": {
-              "type": "string",
-              "description": "Path to the source HTML file (relative to the workspace root)."
-            },
-            "output": {
-              "type": "string",
-              "description": "Output file name written into the output directory.",
-              "default": "index.html"
-            }
-          },
-          "required": ["input"]
-        },
-        {
-          "type": "boolean",
-          "enum": [false],
-          "description": "Set to false to skip writing any index.html."
-        }
-      ]
     },
     "rebuildDelay": {
       "type": "number",
       "default": 2000,
-      "description": "The delay for rebuilding federation artifacts. This allows to have more resources for refreshing your micro frontend in the browser."
+      "description": "Debounce delay (ms) before rebuilding federation artifacts after a file change in watch mode."
     },
     "projectName": {
       "type": "string"
@@ -75,21 +42,6 @@
       "default": true,
       "description": "Will cache the shared externals so they can be re-used between builds",
       "alias": "cache"
-    },
-    "buildNotifications": {
-      "type": "object",
-      "properties": {
-        "enable": {
-          "type": "boolean",
-          "default": true,
-          "description": "Enable build completion notifications for local development. It will send events to notify when the federation build is complete."
-        },
-        "endpoint": {
-          "type": "string",
-          "default": "/@angular-architects/native-federation:build-notifications",
-          "description": "You can override the default endpoint to send build completion events to."
-        }
-      }
     },
     "assets": {
       "type": "array",
